+++ pkg/terminal/starlark_test.go
@@ -2,11 +2,15 @@ package terminal
 
 import (
 	"fmt"
+	"runtime"
 	"strings"
 	"testing"
 )
 
 func TestStarlarkExamples(t *testing.T) {
+	if runtime.GOARCH == "arm64" {
+		t.Skip("test is not valid on ARM64")
+	}
 	withTestTerminal("testvariables2", t, func(term *FakeTerminal) {
 		term.MustExec("continue")
 		t.Run("goroutine_start_line", func(t *testing.T) { testStarlarkExampleGoroutineStartLine(t, term) })
@@ -119,6 +123,9 @@ func testStarlarkMapIteration(t *testing.T, term *FakeTerminal) {
 }
 
 func TestStarlarkVariable(t *testing.T) {
+	if runtime.GOARCH == "arm64" {
+		t.Skip("test is not valid on ARM64")
+	}
 	withTestTerminal("testvariables2", t, func(term *FakeTerminal) {
 		term.MustExec("continue")
 		for _, tc := range []struct{ expr, tgt string }{
