+++ service/test/integration1_test.go
@@ -668,6 +668,9 @@ func Test1ClientServer_FindLocations(t *testing.T) {
 }
 
 func Test1ClientServer_FindLocationsAddr(t *testing.T) {
+	if runtime.GOARCH == "arm64" {
+		t.Skip("test is not valid on ARM64")
+	}
 	withTestClient1("locationsprog2", t, func(c *rpc1.RPCClient) {
 		<-c.Continue()
 
@@ -680,6 +683,9 @@ func Test1ClientServer_FindLocationsAddr(t *testing.T) {
 }
 
 func Test1ClientServer_EvalVariable(t *testing.T) {
+	if runtime.GOARCH == "arm64" {
+		t.Skip("test is not valid on ARM64")
+	}
 	withTestClient1("testvariables", t, func(c *rpc1.RPCClient) {
 		state := <-c.Continue()
 
@@ -699,6 +705,9 @@ func Test1ClientServer_EvalVariable(t *testing.T) {
 }
 
 func Test1ClientServer_SetVariable(t *testing.T) {
+	if runtime.GOARCH == "arm64" {
+		t.Skip("test is not valid on ARM64")
+	}
 	withTestClient1("testvariables", t, func(c *rpc1.RPCClient) {
 		state := <-c.Continue()
 
@@ -724,6 +733,9 @@ func Test1ClientServer_SetVariable(t *testing.T) {
 }
 
 func Test1ClientServer_FullStacktrace(t *testing.T) {
+	if runtime.GOARCH == "arm64" {
+		t.Skip("test is not valid on ARM64")
+	}
 	withTestClient1("goroutinestackprog", t, func(c *rpc1.RPCClient) {
 		_, err := c.CreateBreakpoint(&api.Breakpoint{FunctionName: "main.stacktraceme", Line: -1})
 		assertNoError(err, t, "CreateBreakpoint()")
@@ -796,6 +808,9 @@ func Test1ClientServer_FullStacktrace(t *testing.T) {
 }
 
 func Test1Issue355(t *testing.T) {
+	if runtime.GOARCH == "arm64" {
+		t.Skip("test is not valid on ARM64")
+	}
 	// After the target process has terminated should return an error but not crash
 	withTestClient1("continuetestprog", t, func(c *rpc1.RPCClient) {
 		bp, err := c.CreateBreakpoint(&api.Breakpoint{FunctionName: "main.sayhi", Line: -1})
@@ -854,6 +869,9 @@ func Test1Issue355(t *testing.T) {
 }
 
 func Test1Disasm(t *testing.T) {
+	if runtime.GOARCH == "arm64" {
+		t.Skip("test is not valid on ARM64")
+	}
 	// Tests that disassembling by PC, range, and current PC all yeld similar results
 	// Tests that disassembly by current PC will return a disassembly containing the instruction at PC
 	// Tests that stepping on a calculated CALL instruction will yield a disassembly that contains the
@@ -1022,6 +1040,9 @@ func Test1Issue419(t *testing.T) {
 }
 
 func Test1TypesCommand(t *testing.T) {
+	if runtime.GOARCH == "arm64" {
+		t.Skip("test is not valid on ARM64")
+	}
 	withTestClient1("testvariables2", t, func(c *rpc1.RPCClient) {
 		state := <-c.Continue()
 		assertNoError(state.Err, t, "Continue()")
